+++ sample-apps/react/zoom-clone/src/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from './useChatClient';
+export * from './useSetCall';