+++ packages/styling/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [1.0.4](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-1.0.3...@stream-io/video-styling-1.0.4) (2024-05-23)
+
+
+### Bug Fixes
+
+* popup styling ([#1363](https://github.com/GetStream/stream-video-js/issues/1363)) ([6549d4b](https://github.com/GetStream/stream-video-js/commit/6549d4bd80e198169e8d09920c46dcb487eb071b))
+
 ### [1.0.3](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-1.0.2...@stream-io/video-styling-1.0.3) (2024-05-17)